requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
PyYAML==6.0
deepdiff==6.3.1
jinja2==3.1.2

# CORS Middleware
python-multipart==0.0.6
